+++ multilint/__main__.py
@@ -1,0 +1,10 @@
+# -*- encoding:utf-8 -*-
+from __future__ import (
+    absolute_import, division, print_function, unicode_literals
+)
+
+from . import main
+
+
+if __name__ == '__main__':
+    main()