+++ HISTORY.rst
@@ -6,6 +6,7 @@ Pending
 -------
 
 * New notes here
+* Support ``python -m multilint``
 
 2.0.2 (2016-12-06)
 ------------------